fix development mode in Firefox/Safari/Edge
This refactors the service worker to switch from a module-based approach to an approach understood
by all browsers.

In an ideal world we would use a module-based approach for this file, enabling us to do things like:

  import ... from 'workbox-...';
  const isDev = import.meta.env.DEPLOY_ENV === 'dev'

And in general, that approach works. When the app is built, the build process bundles all of the
imported code and replaces the import metadata with hardcoded values, so the built version of the
service worker works flawlessly in any browser. However, running in local dev mode (i.e. `yarn dev`)
doesn't bundle the code. This means that the browser receives the raw module-based code with `import ...`
and all. Currently, Chrome is the only browser that supports module features for service workers, so
this means that Firefox/Safari/Edge can't be used in dev mode.

The solution this commit switches to is using "classic" mode in VitePWAs dev options and using
`importScripts` in the service worker. We include the workbox library code we need in our codebase
(meh, I know) and also need to keep our workbox plugin code in pure JS files (since the TS can't be
easily compiled on the fly). This approach works well, but requires some extra thought in order for
types to keep working. The workbox types need to be pulled in and applied for each workbox library
(see below) and our workbox plugin code must be typed with JSDoc.

+++ src/lib/data-cache.ts
@@ -90,10 +90,12 @@ export async function clearEntireCache() {
         await caches.delete(key);
     });
 
-    const dbs = await indexedDB.databases();
-    dbs.forEach((db) => {
-        db.name && indexedDB.deleteDatabase(db.name);
-    });
+    if (indexedDB.databases) {
+        const dbs = await indexedDB.databases();
+        dbs.forEach((db) => {
+            db.name && indexedDB.deleteDatabase(db.name);
+        });
+    }
 
     localStorage.clear();
     sessionStorage.clear();
